requirements.txt
coverage==5.5
deprecation==2.1.0
eyed3==0.9.7
filetype==1.2.0
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
pygame==2.1.3
RPi.GPIO==0.7.1
spidev==3.6
ST7789==0.0.4
toml==0.10.2
